Add site module to deploy frontend to a dir
This commit adds a site.pp file which is used to deploy the frontend
to a directory instead of setting up a vhost. When deploying this on
status.o.o we need to put the dashboard in a dir on an existing vhost
instead of deploying a seperate one. This commit adds the class to do
this.

Change-Id: Ic47fbd817a293a257f7a9b164682fc96cce5321e