Enable mod_cache to do API server side caching
This commit adds enabling mod_cache to the api server to enable some
server side caching. We don't need to update the data in real time on
the dashboard and performing some server side caching will improve
the dashboard performance. It'll also decrease the load we'll put on
that poor trove node in rax cloud we're running subunit2sql on.

Change-Id: I61d862603306a36c3ec6cd74d000126155bb29d5