IPC-76 Fix broken acceptance tests
While applying this module downstream we noticed that its acceptance
tests were broken.
During the process of fixing them we noticed that the reason for this
were the changes made in the manifest 'openstack_health::frontend'.
Apparently a new manifest was created to take care of installing the
frontend, 'openstack_health::vhost'.
The first step in order to fix this was adapting the tests to use this
new manifest instead of the former. While doing so we noticed that the
apache vhost template file was not being generated correctly. That
happened because it was binding agains 'openstack_health::frontend',
which now did not have the parameters necessary.
So, in order to make the whole acceptance test suite work, it was
necessary updating the vhost template file as well.

Change-Id: I262eb99a5a8ebf71437ae2a0878d66118fd47faa