Make sure we install the latest requirements
This commit switches the api server from installing the requirements
file with pip install, via the ::python::requirements class, to using
pip install -U explicitly via an exec. We need to use -U to ensure
we're pulling in the latest version of subunit2sql that's valid,
otherwise the version won't be in sync with our database which can
potentially cause issues.

Change-Id: I6797bcf82cc82a8a49441f9abb90bbc95fbd365b