Add build_newrev to the list of ignored keys
A new run metadata key, build_newrev, started showing up in the db
recently. [1] It's not actually something we want to expose, since it
is unique for each job and not something we can group on. This commit
adds it to the list of ignored keys so we don't have to see it
anymore.

[1] http://status.openstack.org/openstack-health/#/?groupKey=build_newrev

Change-Id: Iea4a0babe10cd8d54a31e39dc5c3738570fda126