Stop using a venv for the api server
Previously we were installing the openstack-health api server in a
virtualenv. This was for questionable benefit because we do not actually
need to isolate it from other python environments on the system. This
also had the added disadvantage of greatly complicating our install time
and upgrade story, especially if a new dep was added directly or
indirectly. To address this commit switches the module from using a venv
to directly installing things into system python.

Change-Id: I501711329197eaa9400538272a488909b63b6f0f