Fully root pip path
On initial pip run the bootstrapping fails because pip is not yet in
puppets path (earlier puppet tasks install pip). We can deal with this
by fully rooting the path to pip which should find the binary.

Unfortunately we end up doing this with a exec only if condition to
determine if we should run this under /usr/local/bin/pip or /usr/bin/pip
as there doesn't seem to be a good way to check otherwise.

Change-Id: Ib82f2571a50dad12ab64cfecffdf9d84fae2cf9a